Add long conversion for unemp
Add a conversion to long format for unemployment data as the control change through time with date format in use in this repo.  Add code for export of controls (time invariant and time variant). Change evol_diff into multiplier but either way it may have to be corrected for inflation in some specifications.